Fix downloadProgress being either 0 or 1

+++ musicdao/src/main/java/com/example/musicdao/Track.kt
@@ -56,7 +56,7 @@ class Track(
 
     fun setDownloadProgress(fileProgress: Long, fullSize: Long?): Int {
         val size = fullSize ?: Long.MAX_VALUE
-        val progress: Float = (fileProgress / size) * 100.0f
+        val progress: Double = (fileProgress.toDouble() / size.toDouble()) * 100.0
         progressBar.secondaryProgress = progress.toInt()
         return progress.toInt()
     }
